Updated version
Added info/edit window that can be used differently depending on current user. Two separate popup-windows.

Added two properties to UpdateArticle in MSSQL-repo.
articleID for selecting article to update.
edition because it was added after function was made.

Added enum Edit = 10

CurrentRole is set to Administrator, so that the extended info/edit window is shown